--- pilot_api/application/__init__.py
@@ -1,5 +1,0 @@
-from flask import Flask, request 
-
-app = Flask(__name__)
-
-from application import routes
